Split EnumerableExtensions
