+++ src/FuncSharp/Collections/IEnumerableExtensions_Options.cs
@@ -1,0 +1,61 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace FuncSharp
+{
+    public static partial class IEnumerableExtensions
+    {
+        /// <summary>
+        /// Returns values of the nonempty options.
+        /// </summary>
+        public static IEnumerable<T> Flatten<T>(this IEnumerable<IOption<T>> source)
+        {
+            return source.SelectMany(o => o.ToList());
+        }
+
+        public static IOption<TValue> SafeMax<T, TValue>(this IEnumerable<T> source, Func<T, TValue> selector)
+        {
+            return source.AsNonEmpty().Map(s => s.Max(selector));
+        }
+
+        public static IOption<TValue> SafeMin<T, TValue>(this IEnumerable<T> source, Func<T, TValue> selector)
+        {
+            return source.AsNonEmpty().Map(s => s.Min(selector));
+        }
+
+        /// <summary>
+        /// Returns first value or an empty option.
+        /// </summary>
+        public static IOption<T> FirstOption<T>(this IEnumerable<T> source, Func<T, bool> predicate = null)
+        {
+            var data = source.Where(predicate ?? (t => true)).Take(1).ToList();
+            if (data.Count == 0)
+            {
+                return Option.Empty<T>();
+            }
+            return Option.Valued(data.First());
+        }
+
+        /// <summary>
+        /// Returns last value or an empty option.
+        /// </summary>
+        public static IOption<T> LastOption<T>(this IEnumerable<T> source, Func<T, bool> predicate = null)
+        {
+            return source.Reverse().FirstOption(predicate);
+        }
+
+        /// <summary>
+        /// Returns the only value if the source contains just one value, otherwise an empty option.
+        /// </summary>
+        public static IOption<T> SingleOption<T>(this IEnumerable<T> source, Func<T, bool> predicate = null)
+        {
+            var data = source.Where(predicate ?? (t => true)).Take(2).ToList();
+            if (data.Count == 2)
+            {
+                return Option.Empty<T>();
+            }
+            return data.FirstOption();
+        }
+    }
+}
