+++ src/FuncSharp/Collections/IEnumerableExtensions_PartitionMatch.cs
@@ -1,567 +1,11 @@
 ﻿
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
 
 namespace FuncSharp
 {
-    public static class IEnumerableExtensions
+    public static partial class IEnumerableExtensions
     {
-        #region Generic
-
-        /// <summary>
-        /// Returns a ToList() juts of type IReadOnlyList.
-        /// </summary>
-        public static IReadOnlyList<T> ToReadOnlyList<T>(this IEnumerable<T> e)
-        {
-            return e.ToList().AsReadOnly();
-        }
-
-        /// <summary>
-        /// Returns the IEnumerable in case it is a ReadOnlyList or creates a new ReadOnlyList from it.
-        /// </summary>
-        [DebuggerStepThrough]
-        public static IReadOnlyList<T> AsReadOnlyList<T>(this IEnumerable<T> source)
-        {
-            return source as IReadOnlyList<T> ?? source.ToList();
-        }
-
-        /// <summary>
-        /// Returns the IEnumerable in case it is a List or creates a new List from it.
-        /// </summary>
-        [DebuggerStepThrough]
-        public static List<T> AsList<T>(this IEnumerable<T> source)
-        {
-            return source as List<T> ?? source.ToList();
-        }
-
-        /// <summary>
-        /// Returns the IEnumerable in case it is a Array or creates a new Array from it.
-        /// </summary>
-        [DebuggerStepThrough]
-        public static T[] AsArray<T>(this IEnumerable<T> source)
-        {
-            return source as T[] ?? source.ToArray();
-        }
-
-        /// <summary>
-        /// Returns all the items inside all the collections combined into 1 IEnumerable.
-        /// </summary>
-        public static IEnumerable<T> Flatten<T>(this IEnumerable<IEnumerable<T>> e)
-        {
-            return e.SelectMany(i => i);
-        }
-
-        public static IEnumerable<T> Except<T>(this IEnumerable<T> e, params T[] excludedItems)
-        {
-            return e.Except(excludedItems.AsEnumerable());
-        }
-
-        public static IEnumerable<T> Except<T>(this IEnumerable<T> e, params IEnumerable<T>[] others)
-        {
-            return Enumerable.Except(e, others.Flatten());
-        }
-
-        public static IEnumerable<T> ExceptNulls<T>(this IEnumerable<T?> e)
-            where T : struct
-        {
-            return e.Where(v => v.HasValue).Select(v => v.Value);
-        }
-
-        public static IEnumerable<T> ExceptNulls<T>(this IEnumerable<T> e)
-            where T : class
-        {
-            return e.Where(v => v is not null);
-        }
-
-        public static bool NonEmpty<T>(this IEnumerable<T> e)
-        {
-            return e is not null && e.Any();
-        }
-
-        public static bool IsEmpty<T>(this IEnumerable<T> e)
-        {
-            return !e.NonEmpty();
-        }
-
-        public static bool IsMultiple<T>(this IEnumerable<T> e)
-        {
-            return e is not null && e.Take(2).Count().SafeEquals(2);
-        }
-
-        public static bool IsSingle<T>(this IEnumerable<T> e)
-        {
-            return e is not null && e.Take(2).Count().SafeEquals(1);
-        }
-
-        public static T Second<T>(this IEnumerable<T> e)
-        {
-            return e.ElementAt(1);
-        }
-
-        public static T Third<T>(this IEnumerable<T> e)
-        {
-            return e.ElementAt(2);
-        }
-
-        public static T Fourth<T>(this IEnumerable<T> e)
-        {
-            return e.ElementAt(3);
-        }
-
-        public static T Fifth<T>(this IEnumerable<T> e)
-        {
-            return e.ElementAt(4);
-        }
-
-        public static IEnumerable<T> Concat<T>(this IEnumerable<T> first, params T[] items)
-        {
-            return Enumerable.Concat(first, items);
-        }
-
-        public static IEnumerable<T> Concat<T>(this IEnumerable<T> first, params IEnumerable<T>[] others)
-        {
-            return Enumerable.Concat(first, others.Flatten());
-        }
-
-        public static IEnumerable<T> SafeConcat<T>(this IEnumerable<T> first, params T[] items)
-        {
-            return Enumerable.Concat(first ?? Enumerable.Empty<T>(), items);
-        }
-
-        public static IEnumerable<T> SafeConcat<T>(this IEnumerable<T> first, params IEnumerable<T>[] others)
-        {
-            return Enumerable.Concat(first ?? Enumerable.Empty<T>(), others.SelectMany(o => o ?? Enumerable.Empty<T>()));
-        }
-
-        /// <summary>
-        /// Orders the values using the specified less function in the specified order.
-        /// </summary>
-        public static IEnumerable<T> Order<T>(this IEnumerable<T> values, Func<T, T, bool> less, Ordering ordering = Ordering.Ascending)
-        {
-            var comparer = new Comparer<T>(less, ordering);
-            return values.OrderBy(v => v, comparer);
-        }
-
-        /// <summary>
-        /// Aggregates the exceptions into an AggregateException. If there is a single exception, returns it directly.
-        /// </summary>
-        public static IOption<Exception> Aggregate(this IEnumerable<Exception> source)
-        {
-            var exceptions = source.ToList();
-            switch (exceptions.Count)
-            {
-                case 0: return Option<Exception>.Empty;
-                case 1: return Option.Valued(exceptions[0]);
-                default: return Option.Valued(new AggregateException(exceptions));
-            }
-        }
-
-        #endregion Generic
-
-        #region Numeric
-
-        public static PositiveInt Sum(this INonEmptyEnumerable<PositiveInt> values)
-        {
-            return PositiveInt.CreateUnsafe(values.Sum(v => v.Value));
-        }
-
-        public static NonPositiveInt Sum(this IEnumerable<NonPositiveInt> values)
-        {
-            return values.Aggregate(0.AsUnsafeNonPositive(), (a, b) => a + b);
-        }
-
-        public static NonNegativeInt Sum(this IEnumerable<NonNegativeInt> values)
-        {
-            return values.Aggregate(NonNegativeInt.Zero, (a, b) => a + b);
-        }
-
-        #endregion Numeric
-
-        #region NonEmpty
-
-        /// <summary>
-        /// Returns a nonEmptyEnumerable in case the collection is nonempty. Otherwise returns empty option.
-        /// </summary>
-        [DebuggerStepThrough]
-        public static IOption<INonEmptyEnumerable<T>> AsNonEmpty<T>(this IEnumerable<T> source)
-        {
-            return source is null
-                ? Option.Empty<INonEmptyEnumerable<T>>()
-                : NonEmptyEnumerable.Create(source);
-        }
-
-        public static INonEmptyEnumerable<T> Flatten<T>(this INonEmptyEnumerable<INonEmptyEnumerable<T>> source)
-        {
-            return NonEmptyEnumerable<T>.Create(source.Head.Head, source.Head.Tail.Concat(source.Tail.Flatten()).ToArray());
-        }
-
-        public static INonEmptyEnumerable<T> Concat<T>(this T e, params IEnumerable<T>[] others)
-        {
-            return NonEmptyEnumerable.Create(e, others.Flatten().ToArray());
-        }
-
-        public static INonEmptyEnumerable<T> Concat<T>(this INonEmptyEnumerable<T> source, params T[] items)
-        {
-            return NonEmptyEnumerable.Create(source.Head, source.Tail.Concat(items).ToArray());
-        }
-
-        public static INonEmptyEnumerable<T> Concat<T>(this INonEmptyEnumerable<T> source, params IEnumerable<T>[] items)
-        {
-            return NonEmptyEnumerable.Create(source.Head, source.Tail.Concat(items).ToArray());
-        }
-
-        #endregion NonEmpty
-
-        #region Options
-
-        /// <summary>
-        /// Returns values of the nonempty options.
-        /// </summary>
-        public static IEnumerable<T> Flatten<T>(this IEnumerable<IOption<T>> source)
-        {
-            return source.SelectMany(o => o.ToList());
-        }
-
-        public static IOption<TValue> SafeMax<T, TValue>(this IEnumerable<T> source, Func<T, TValue> selector)
-        {
-            return source.AsNonEmpty().Map(s => s.Max(selector));
-        }
-
-        public static IOption<TValue> SafeMin<T, TValue>(this IEnumerable<T> source, Func<T, TValue> selector)
-        {
-            return source.AsNonEmpty().Map(s => s.Min(selector));
-        }
-
-        /// <summary>
-        /// Returns first value or an empty option.
-        /// </summary>
-        public static IOption<T> FirstOption<T>(this IEnumerable<T> source, Func<T, bool> predicate = null)
-        {
-            var data = source.Where(predicate ?? (t => true)).Take(1).ToList();
-            if (data.Count == 0)
-            {
-                return Option.Empty<T>();
-            }
-            return Option.Valued(data.First());
-        }
-
-        /// <summary>
-        /// Returns last value or an empty option.
-        /// </summary>
-        public static IOption<T> LastOption<T>(this IEnumerable<T> source, Func<T, bool> predicate = null)
-        {
-            return source.Reverse().FirstOption(predicate);
-        }
-
-        /// <summary>
-        /// Returns the only value if the source contains just one value, otherwise an empty option.
-        /// </summary>
-        public static IOption<T> SingleOption<T>(this IEnumerable<T> source, Func<T, bool> predicate = null)
-        {
-            var data = source.Where(predicate ?? (t => true)).Take(2).ToList();
-            if (data.Count == 2)
-            {
-                return Option.Empty<T>();
-            }
-            return data.FirstOption();
-        }
-
-        #endregion Options
-
-        #region DataCubes
-
-        /// <summary>
-        /// Converts the source to a new 1-dimensional data cube and aggregates the values in case of conflicting positions.
-        /// </summary>
-        public static DataCube1<P1, TResult> ToDataCube<T, P1, TValue, TResult>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, TValue> value,
-            Func<TValue, TResult> initialization,
-            Func<TResult, TValue, TResult> aggregation)
-        {
-            var dataCube = new DataCube1<P1, TResult>();
-            foreach (var v in source)
-            {
-                dataCube.SetOrElseUpdate<TValue>(p1(v), value(v), initialization, aggregation);
-            }
-            return dataCube;
-        }
-
-        /// <summary>
-        /// Converts the source to a new 2-dimensional data cube and aggregates the values in case of conflicting positions.
-        /// </summary>
-        public static DataCube2<P1, P2, TResult> ToDataCube<T, P1, P2, TValue, TResult>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, TValue> value,
-            Func<TValue, TResult> initialization,
-            Func<TResult, TValue, TResult> aggregation)
-        {
-            var dataCube = new DataCube2<P1, P2, TResult>();
-            foreach (var v in source)
-            {
-                dataCube.SetOrElseUpdate<TValue>(p1(v), p2(v), value(v), initialization, aggregation);
-            }
-            return dataCube;
-        }
-
-        /// <summary>
-        /// Converts the source to a new 3-dimensional data cube and aggregates the values in case of conflicting positions.
-        /// </summary>
-        public static DataCube3<P1, P2, P3, TResult> ToDataCube<T, P1, P2, P3, TValue, TResult>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, TValue> value,
-            Func<TValue, TResult> initialization,
-            Func<TResult, TValue, TResult> aggregation)
-        {
-            var dataCube = new DataCube3<P1, P2, P3, TResult>();
-            foreach (var v in source)
-            {
-                dataCube.SetOrElseUpdate<TValue>(p1(v), p2(v), p3(v), value(v), initialization, aggregation);
-            }
-            return dataCube;
-        }
-
-        /// <summary>
-        /// Converts the source to a new 4-dimensional data cube and aggregates the values in case of conflicting positions.
-        /// </summary>
-        public static DataCube4<P1, P2, P3, P4, TResult> ToDataCube<T, P1, P2, P3, P4, TValue, TResult>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, P4> p4,
-            Func<T, TValue> value,
-            Func<TValue, TResult> initialization,
-            Func<TResult, TValue, TResult> aggregation)
-        {
-            var dataCube = new DataCube4<P1, P2, P3, P4, TResult>();
-            foreach (var v in source)
-            {
-                dataCube.SetOrElseUpdate<TValue>(p1(v), p2(v), p3(v), p4(v), value(v), initialization, aggregation);
-            }
-            return dataCube;
-        }
-
-        /// <summary>
-        /// Converts the source to a new 5-dimensional data cube and aggregates the values in case of conflicting positions.
-        /// </summary>
-        public static DataCube5<P1, P2, P3, P4, P5, TResult> ToDataCube<T, P1, P2, P3, P4, P5, TValue, TResult>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, P4> p4,
-            Func<T, P5> p5,
-            Func<T, TValue> value,
-            Func<TValue, TResult> initialization,
-            Func<TResult, TValue, TResult> aggregation)
-        {
-            var dataCube = new DataCube5<P1, P2, P3, P4, P5, TResult>();
-            foreach (var v in source)
-            {
-                dataCube.SetOrElseUpdate<TValue>(p1(v), p2(v), p3(v), p4(v), p5(v), value(v), initialization, aggregation);
-            }
-            return dataCube;
-        }
-
-        /// <summary>
-        /// Converts the source to a new 6-dimensional data cube and aggregates the values in case of conflicting positions.
-        /// </summary>
-        public static DataCube6<P1, P2, P3, P4, P5, P6, TResult> ToDataCube<T, P1, P2, P3, P4, P5, P6, TValue, TResult>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, P4> p4,
-            Func<T, P5> p5,
-            Func<T, P6> p6,
-            Func<T, TValue> value,
-            Func<TValue, TResult> initialization,
-            Func<TResult, TValue, TResult> aggregation)
-        {
-            var dataCube = new DataCube6<P1, P2, P3, P4, P5, P6, TResult>();
-            foreach (var v in source)
-            {
-                dataCube.SetOrElseUpdate<TValue>(p1(v), p2(v), p3(v), p4(v), p5(v), p6(v), value(v), initialization, aggregation);
-            }
-            return dataCube;
-        }
-
-
-        /// <summary>
-        /// Converts the source to a new 1-dimensional data cube.
-        /// </summary>
-        public static DataCube1<P1, TValue> ToDataCube<T, P1, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, TValue, TValue>(source, p1, value, a => a, aggregation: (a, b) => throw new ArgumentException("An item with the same key has already been added."));
-        }
-
-        /// <summary>
-        /// Converts the source to a new 2-dimensional data cube.
-        /// </summary>
-        public static DataCube2<P1, P2, TValue> ToDataCube<T, P1, P2, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, P2, TValue, TValue>(source, p1, p2, value, a => a, aggregation: (a, b) => throw new ArgumentException("An item with the same key has already been added."));
-        }
-
-        /// <summary>
-        /// Converts the source to a new 3-dimensional data cube.
-        /// </summary>
-        public static DataCube3<P1, P2, P3, TValue> ToDataCube<T, P1, P2, P3, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, P2, P3, TValue, TValue>(source, p1, p2, p3, value, a => a, aggregation: (a, b) => throw new ArgumentException("An item with the same key has already been added."));
-        }
-
-        /// <summary>
-        /// Converts the source to a new 4-dimensional data cube.
-        /// </summary>
-        public static DataCube4<P1, P2, P3, P4, TValue> ToDataCube<T, P1, P2, P3, P4, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, P4> p4,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, P2, P3, P4, TValue, TValue>(source, p1, p2, p3, p4, value, a => a, aggregation: (a, b) => throw new ArgumentException("An item with the same key has already been added."));
-        }
-
-        /// <summary>
-        /// Converts the source to a new 5-dimensional data cube.
-        /// </summary>
-        public static DataCube5<P1, P2, P3, P4, P5, TValue> ToDataCube<T, P1, P2, P3, P4, P5, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, P4> p4,
-            Func<T, P5> p5,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, P2, P3, P4, P5, TValue, TValue>(source, p1, p2, p3, p4, p5, value, a => a, aggregation: (a, b) => throw new ArgumentException("An item with the same key has already been added."));
-        }
-
-        /// <summary>
-        /// Converts the source to a new 6-dimensional data cube.
-        /// </summary>
-        public static DataCube6<P1, P2, P3, P4, P5, P6, TValue> ToDataCube<T, P1, P2, P3, P4, P5, P6, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, P4> p4,
-            Func<T, P5> p5,
-            Func<T, P6> p6,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, P2, P3, P4, P5, P6, TValue, TValue>(source, p1, p2, p3, p4, p5, p6, value, a => a, aggregation: (a, b) => throw new ArgumentException("An item with the same key has already been added."));
-        }
-
-        /// <summary>
-        /// Converts the source to a new 1-dimensional data cube, in case of collisions, it aggregates the values to a collection.
-        /// </summary>
-        public static DataCube1<P1, IEnumerable<TValue>> ToCollectionDataCube<T, P1, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, TValue, IEnumerable<TValue>>(source, p1, value, a => Enumerable.Repeat(a, 1), (a, b) => a.Concat(new [] { b }));
-        }
-
-        /// <summary>
-        /// Converts the source to a new 2-dimensional data cube, in case of collisions, it aggregates the values to a collection.
-        /// </summary>
-        public static DataCube2<P1, P2, IEnumerable<TValue>> ToCollectionDataCube<T, P1, P2, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, P2, TValue, IEnumerable<TValue>>(source, p1, p2, value, a => Enumerable.Repeat(a, 1), (a, b) => a.Concat(new [] { b }));
-        }
-
-        /// <summary>
-        /// Converts the source to a new 3-dimensional data cube, in case of collisions, it aggregates the values to a collection.
-        /// </summary>
-        public static DataCube3<P1, P2, P3, IEnumerable<TValue>> ToCollectionDataCube<T, P1, P2, P3, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, P2, P3, TValue, IEnumerable<TValue>>(source, p1, p2, p3, value, a => Enumerable.Repeat(a, 1), (a, b) => a.Concat(new [] { b }));
-        }
-
-        /// <summary>
-        /// Converts the source to a new 4-dimensional data cube, in case of collisions, it aggregates the values to a collection.
-        /// </summary>
-        public static DataCube4<P1, P2, P3, P4, IEnumerable<TValue>> ToCollectionDataCube<T, P1, P2, P3, P4, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, P4> p4,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, P2, P3, P4, TValue, IEnumerable<TValue>>(source, p1, p2, p3, p4, value, a => Enumerable.Repeat(a, 1), (a, b) => a.Concat(new [] { b }));
-        }
-
-        /// <summary>
-        /// Converts the source to a new 5-dimensional data cube, in case of collisions, it aggregates the values to a collection.
-        /// </summary>
-        public static DataCube5<P1, P2, P3, P4, P5, IEnumerable<TValue>> ToCollectionDataCube<T, P1, P2, P3, P4, P5, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, P4> p4,
-            Func<T, P5> p5,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, P2, P3, P4, P5, TValue, IEnumerable<TValue>>(source, p1, p2, p3, p4, p5, value, a => Enumerable.Repeat(a, 1), (a, b) => a.Concat(new [] { b }));
-        }
-
-        /// <summary>
-        /// Converts the source to a new 6-dimensional data cube, in case of collisions, it aggregates the values to a collection.
-        /// </summary>
-        public static DataCube6<P1, P2, P3, P4, P5, P6, IEnumerable<TValue>> ToCollectionDataCube<T, P1, P2, P3, P4, P5, P6, TValue>(
-            this IEnumerable<T> source,
-            Func<T, P1> p1,
-            Func<T, P2> p2,
-            Func<T, P3> p3,
-            Func<T, P4> p4,
-            Func<T, P5> p5,
-            Func<T, P6> p6,
-            Func<T, TValue> value)
-        {
-            return ToDataCube<T, P1, P2, P3, P4, P5, P6, TValue, IEnumerable<TValue>>(source, p1, p2, p3, p4, p5, p6, value, a => Enumerable.Repeat(a, 1), (a, b) => a.Concat(new [] { b }));
-        }
-
-        #endregion DataCubes
-
-        #region PartitionMatch
-
 
         /// <summary>
         /// For each partition (collection of n-th coproduct elements), invokes the specified function.
@@ -2402,8 +1846,5 @@ namespace FuncSharp
 
             return result;
         }
-
-        #endregion PartitionMatch
-
     }
 }
